sec(pg): cast make_interval() integer args ::int (pg17 runtime failures)
PostgreSQL's make_interval() types every argument as `int` (int4) EXCEPT
`secs` (double precision). sqlx sends a bound parameter with the OID derived
from the Rust type, so binding an i64 (int8) or f64 (float8) to
`make_interval(hours => $N)` resolves to a NON-EXISTENT overload and fails at
request time on the default pgvector pg17 deploy (and pg16):

  ERROR: function make_interval(hours => bigint) does not exist

int8→int4 / float8→int4 are assignment casts, not implicit, so pg's overload
resolution won't apply them. This compiles clean under `cargo check` and only
trips when the query runs — empirically confirmed against pg17 in this change.

Real bugs found (500 at request time on encrypted/default deploys):
- talos-retry-intelligence + talos-cost-attribution bound `hours as f64` into
  `make_interval(hours => $N)` — float8, fails. (The `as f64` was a misguided
  "it wants a number" cast; cleaned to `.bind(hours)`.)
- SecretsManager::list_secret_access_log takes `hours: f64` → same failure.
The crash-recovery claim hit the i64 variant earlier (#51, fixed then).

Fix: cast the parameter in SQL — `make_interval(hours => $N::int)` — which
coerces any numeric bind (i32/i64/f64) to int4. Applied uniformly to all 35
non-`secs` sites across 10 files, so the invariant holds regardless of the
bound Rust type (and future i32→i64 param changes stay safe). `secs =>`
(node-cache TTL) is left alone — it's double precision and accepts int/float.

Frozen by structural lint check 27: flags any
`make_interval((mins|hours|days|weeks|months|years) => $N)` lacking `::int`;
`secs` exempt; opt-out `// allow-make-interval-no-cast: <reason>`.
Sensitivity-tested (flags a reverted site; clean on the fixed tree).

Note: the sibling `INTERVAL '1 day' * $N` retention form (cleanup_* fns) is
unaffected — `interval * bigint` is valid, so i64 binds work there.

Verified: cargo check --workspace clean; all 27 structural lints pass; touched-
crate test suites green (retry-intelligence, cost-attribution, secrets-manager
35, analytics, advanced-repository); make_interval type resolution confirmed
empirically on a live pg17 container.

+++ controller/src/main.rs
@@ -1671,7 +1671,7 @@ async fn main() -> anyhow::Result<()> {
             // override path matches the env-side hardening from MCP-643.
             // Pre-fix a `system_settings.value = 0` row took
             // db_days.unwrap_or(env_days) → 0, then bound 0 into
-            // `make_interval(days => $1)` below — "completed_at < NOW() -
+            // `make_interval(days => $1::int)` below — "completed_at < NOW() -
             // 0 days" matches every completed/failed/cancelled execution
             // ever, archiving the entire table at the next daily tick.
             // Negative DB values would have the same effect (Postgres
@@ -1721,7 +1721,7 @@ async fn main() -> anyhow::Result<()> {
                 "WITH archived AS (
                     DELETE FROM workflow_executions
                     WHERE status IN ('completed', 'failed', 'cancelled')
-                    AND completed_at < NOW() - make_interval(days => $1)
+                    AND completed_at < NOW() - make_interval(days => $1::int)
                     AND is_pinned = false
                     RETURNING *
                 )
@@ -4411,7 +4411,7 @@ async fn main() -> anyhow::Result<()> {
                         "UPDATE workflow_executions SET status = 'failed', \
                          error_message = 'Auto-cleaned: execution stale (running > configured threshold)', \
                          completed_at = NOW() \
-                         WHERE status IN ('running') AND status != 'queued' AND started_at < NOW() - make_interval(mins => $1)"
+                         WHERE status IN ('running') AND status != 'queued' AND started_at < NOW() - make_interval(mins => $1::int)"
                     ).bind(stale_minutes).execute(&cleanup_pool).await;
                     if let Ok(r) = result {
                         if r.rows_affected() > 0 {
